+++ CHANGELOG.md
@@ -114,6 +114,8 @@ version 0.2.0, thus covering only a fraction of the project’s history.
 - A more categorical fix for `dmote-keycap` parameter support, achieved by
   migrating that library’s parsing logic into the library itself.
   See version 0.5.1.
+- Better placement of heat-set inserts. More careful interaction with
+  a plate offset and with the bottom plate.
 - Slightly improved user feedback when configuration files contain structural
   problems.